Moved PSBool to Plutarch.Bool
