+++ Plutarch/Bool.hs
@@ -15,6 +15,16 @@ module Plutarch.Bool (
   pand,
   pand',
   por',
+  PSBool (..),
+  pstrue,
+  psfalse,
+  psif,
+  psif',
+  psnot,
+  psand,
+  psand',
+  psor,
+  psor',
 ) where
 
 import Data.List.NonEmpty (nonEmpty)
@@ -29,24 +39,14 @@ import Generics.SOP (
   ccompare_NS,
   hcliftA2,
  )
-import Plutarch.Internal (
-  PDelayed,
-  S,
-  Term,
-  pdelay,
-  pforce,
-  phoistAcyclic,
-  plet,
-  (#),
-  (#$),
-  (:-->),
- )
+import Plutarch.Internal (PDelayed, PType, S, Term, pdelay, pforce, phoistAcyclic, plet, (#), (#$), (:-->))
 import Plutarch.Internal.Generic (PCode, PGeneric, gpfrom)
 import Plutarch.Internal.Other (
   pto,
  )
 import Plutarch.Internal.PLam (plam)
 import Plutarch.Internal.PlutusType (PInner, PlutusType, pcon, pcon', pmatch, pmatch')
+import Plutarch.Internal.Quantification (PForall (PForall))
 import Plutarch.Lift (
   DerivePConstantDirect (DerivePConstantDirect),
   PConstantDecl,
@@ -182,3 +182,57 @@ eqProd p1 p2 =
     eqTerm :: forall s a. PEq a => Term s a -> Term s a -> K (Term s PBool) a
     eqTerm a b =
       K $ a #== b
+
+-- | 'PInner' of 'PSBool'.
+newtype PSBoolRaw (a :: PType) (s :: S) = PSBoolRaw (Term s (a :--> a :--> a))
+
+instance PlutusType (PSBoolRaw a) where
+  type PInner (PSBoolRaw a) = a :--> a :--> a
+  pcon' (PSBoolRaw x) = x
+  pmatch' x f = f (PSBoolRaw x)
+
+-- | Scott-encoded bool with strict 'pmatch'.
+data PSBool (s :: S)
+  = PSTrue
+  | PSFalse
+
+instance PlutusType PSBool where
+  type PInner PSBool = PForall PSBoolRaw
+  pcon' PSTrue = pcon $ PForall $ pcon $ PSBoolRaw $ plam $ \a _ -> a
+  pcon' PSFalse = pcon $ PForall $ pcon $ PSBoolRaw $ plam $ \_ b -> b
+  pmatch' x' f =
+    pmatch x' $ \(PForall raw) ->
+      pmatch raw $ \(PSBoolRaw x) ->
+        x # f PSTrue # f PSFalse
+
+pstrue :: forall (s :: S). Term s PSBool
+pstrue = pcon PSTrue
+
+psfalse :: forall (s :: S). Term s PSBool
+psfalse = pcon PSFalse
+
+-- | Strict @if@ on Scott-encoded bool.
+psif' :: forall (s :: S) (a :: PType). Term s PSBool -> Term s a -> Term s a -> Term s a
+psif' b t f = pmatch b \case
+  PSTrue -> t
+  PSFalse -> f
+
+-- | Lazy @if@ on Scott-encoded bool.
+psif :: forall (s :: S) (a :: PType). Term s PSBool -> Term s a -> Term s a -> Term s a
+psif b t f = pforce $ psif' b (pdelay t) (pdelay f)
+
+-- | @not@ on Scott-encoded bool.
+psnot :: forall (s :: S). Term s PSBool -> Term s PSBool
+psnot b = psif' b psfalse pstrue
+
+psand' :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool
+psand' a b = psif' a b psfalse
+
+psand :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool
+psand a b = psif a b psfalse
+
+psor' :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool
+psor' a b = psif' a pstrue b
+
+psor :: forall (s :: S). Term s PSBool -> Term s PSBool -> Term s PSBool
+psor a b = psif a pstrue b
